Add isUserLogin() to the all api actions

+++ mpgk-react-frontend/src/actions/activityActions.js
@@ -7,19 +7,22 @@ import {
 	UPDATE_ACTIVITY,
 	GET_ACTIVITIES_BY_CONNECTION
 } from "./types";
+import isUserLogin from "../securityUtils/isUserLogin";
 
 export const addActivity = (activity) => async (dispatch) => {
 	try {
-		const res = await axios.post("/api/activities", activity);
-		dispatch({
-			type: ADD_ACTIVITY,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.post("/api/activities", activity);
+			dispatch({
+				type: ADD_ACTIVITY,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -29,28 +32,34 @@ export const addActivity = (activity) => async (dispatch) => {
 };
 
 export const getActivities = () => async (dispatch) => {
-	const res = await axios.get("/api/activities");
-	dispatch({
-		type: GET_ACTIVITIES,
-		payload: res.data
-	});
+	if (isUserLogin()) {
+		const res = await axios.get("/api/activities");
+		dispatch({
+			type: GET_ACTIVITIES,
+			payload: res.data
+		});
+	}
 };
-
 
 export const updateActivity = (activityId, updateActivity) => async (
 	dispatch
 ) => {
 	try {
-		const res = await axios.put(`/api/activities/${activityId}`, updateActivity);
-		dispatch({
-			type: UPDATE_ACTIVITY,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.put(
+				`/api/activities/${activityId}`,
+				updateActivity
+			);
+			dispatch({
+				type: UPDATE_ACTIVITY,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -60,12 +69,14 @@ export const updateActivity = (activityId, updateActivity) => async (
 };
 
 export const deleteActivity = (activityId) => async (dispatch) => {
-	if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie czynności")) {
-		await axios.delete(`/api/activities/${activityId}`);
-		dispatch({
-			type: DELETE_ACTIVITY,
-			payload: activityId
-		});
+	if (isUserLogin()) {
+		if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie czynności")) {
+			await axios.delete(`/api/activities/${activityId}`);
+			dispatch({
+				type: DELETE_ACTIVITY,
+				payload: activityId
+			});
+		}
 	}
 };
 
@@ -73,11 +84,13 @@ export const getActivitiesByConnection = (connectionId, history) => async (
 	dispach
 ) => {
 	try {
-		const res = await axios.get(`/api/activities/list/${connectionId}`);
-		dispach({
-			type: GET_ACTIVITIES_BY_CONNECTION,
-			payload: res.data
-		});
+		if (isUserLogin()) {
+			const res = await axios.get(`/api/activities/list/${connectionId}`);
+			dispach({
+				type: GET_ACTIVITIES_BY_CONNECTION,
+				payload: res.data
+			});
+		}
 	} catch (error) {
 		history.push("/");
 	}
