+++ mpgk-react-frontend/src/actions/connectionActions.js
@@ -11,16 +11,18 @@ import isUserLogin from "../securityUtils/isUserLogin";
 
 export const addConnection = (connection) => async (dispatch) => {
 	try {
-		const res = await axios.post("/api/connections", connection);
-		dispatch({
-			type: ADD_CONNECTION,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.post("/api/connections", connection);
+			dispatch({
+				type: ADD_CONNECTION,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -30,30 +32,34 @@ export const addConnection = (connection) => async (dispatch) => {
 };
 
 export const getConnections = () => async (dispach) => {
-	const res = await axios.get("/api/connections/");
-	dispach({
-		type: GET_CONNECTIONS,
-		payload: res.data
-	});
+	if (isUserLogin()) {
+		const res = await axios.get("/api/connections/");
+		dispach({
+			type: GET_CONNECTIONS,
+			payload: res.data
+		});
+	}
 };
 
 export const updateConnection = (connectionId, updatedConnection) => async (
 	dispatch
 ) => {
 	try {
-		const res = await axios.put(
-			`/api/connections/${connectionId}`,
-			updatedConnection
-		);
-		dispatch({
-			type: UPDATE_CONNECTION,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.put(
+				`/api/connections/${connectionId}`,
+				updatedConnection
+			);
+			dispatch({
+				type: UPDATE_CONNECTION,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -63,12 +69,14 @@ export const updateConnection = (connectionId, updatedConnection) => async (
 };
 
 export const deleteConnection = (connectionId) => async (dispach) => {
-	if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie powiązania")) {
-		await axios.delete(`/api/connections/${connectionId}`);
-		dispach({
-			type: DELETE_CONNECTION,
-			payload: connectionId
-		});
+	if (isUserLogin()) {
+		if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie powiązania")) {
+			await axios.delete(`/api/connections/${connectionId}`);
+			dispach({
+				type: DELETE_CONNECTION,
+				payload: connectionId
+			});
+		}
 	}
 };
