+++ mpgk-react-frontend/src/actions/fluidActions.js
@@ -6,19 +6,22 @@ import {
 	ADD_FLUID,
 	UPDATE_FLUID
 } from "./types";
+import isUserLogin from "../securityUtils/isUserLogin";
 
 export const addFluid = (fluid) => async (dispatch) => {
 	try {
-		const res = await axios.post("/api/fluids", fluid);
-		dispatch({
-			type: ADD_FLUID,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.post("/api/fluids", fluid);
+			dispatch({
+				type: ADD_FLUID,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -28,25 +31,29 @@ export const addFluid = (fluid) => async (dispatch) => {
 };
 
 export const getFluids = () => async (dispatch) => {
-	const res = await axios.get("/api/fluids");
-	dispatch({
-		type: GET_FLUIDS,
-		payload: res.data
-	});
+	if (isUserLogin()) {
+		const res = await axios.get("/api/fluids");
+		dispatch({
+			type: GET_FLUIDS,
+			payload: res.data
+		});
+	}
 };
 
 export const updateFluid = (fluidId, updatedFluid) => async (dispatch) => {
 	try {
-		const res = await axios.put(`/api/fluids/${fluidId}`, updatedFluid);
-		dispatch({
-			type: UPDATE_FLUID,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.put(`/api/fluids/${fluidId}`, updatedFluid);
+			dispatch({
+				type: UPDATE_FLUID,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -56,12 +63,14 @@ export const updateFluid = (fluidId, updatedFluid) => async (dispatch) => {
 };
 
 export const deleteFluid = (fluidId) => async (dispatch) => {
-	if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie płynu")) {
-		await axios.delete(`/api/fluids/${fluidId}`);
-		dispatch({
-			type: DELETE_FLUID,
-			payload: fluidId
-		});
+	if (isUserLogin()) {
+		if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie płynu")) {
+			await axios.delete(`/api/fluids/${fluidId}`);
+			dispatch({
+				type: DELETE_FLUID,
+				payload: fluidId
+			});
+		}
 	}
 };
