+++ mpgk-react-frontend/src/actions/fluidPlaceActions.js
@@ -6,19 +6,22 @@ import {
 	ADD_FLUID_PLACE,
 	UPDATE_FLUID_PLACE
 } from "./types";
+import isUserLogin from "../securityUtils/isUserLogin";
 
 export const addFluidPlace = (fludidPlace) => async (dispatch) => {
 	try {
-		const res = await axios.post("/api/fluid-places", fludidPlace);
-		dispatch({
-			type: ADD_FLUID_PLACE,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.post("/api/fluid-places", fludidPlace);
+			dispatch({
+				type: ADD_FLUID_PLACE,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -28,30 +31,34 @@ export const addFluidPlace = (fludidPlace) => async (dispatch) => {
 };
 
 export const getFluidPlaces = () => async (dispatch) => {
-	const res = await axios.get("/api/fluid-places");
-	dispatch({
-		type: GET_FLUID_PLACES,
-		payload: res.data
-	});
+	if (isUserLogin()) {
+		const res = await axios.get("/api/fluid-places");
+		dispatch({
+			type: GET_FLUID_PLACES,
+			payload: res.data
+		});
+	}
 };
 
 export const updateFluidPlace = (placeId, updatedFluidPlace) => async (
 	dispatch
 ) => {
 	try {
-		const res = await axios.put(
-			`/api/fluid-places/${placeId}`,
-			updatedFluidPlace
-		);
-		dispatch({
-			type: UPDATE_FLUID_PLACE,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.put(
+				`/api/fluid-places/${placeId}`,
+				updatedFluidPlace
+			);
+			dispatch({
+				type: UPDATE_FLUID_PLACE,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -61,16 +68,18 @@ export const updateFluidPlace = (placeId, updatedFluidPlace) => async (
 };
 
 export const deleteFluidPlace = (placeId) => async (dispatch) => {
-	if (
-		window.confirm(
-			"Czy jesteś pewny? Spowoduje to usunięcie miejsca dodania płynu"
-		)
-	) {
-		await axios.delete(`/api/fluid-places/${placeId}`);
-		dispatch({
-			type: DELETE_FLUID_PLACE,
-			payload: placeId
-		});
+	if (isUserLogin()) {
+		if (
+			window.confirm(
+				"Czy jesteś pewny? Spowoduje to usunięcie miejsca dodania płynu"
+			)
+		) {
+			await axios.delete(`/api/fluid-places/${placeId}`);
+			dispatch({
+				type: DELETE_FLUID_PLACE,
+				payload: placeId
+			});
+		}
 	}
 };
