+++ mpgk-react-frontend/src/actions/fluidRegistryActions.js
@@ -6,19 +6,22 @@ import {
 	ADD_FLUID_REGISTRY,
 	UPDATE_FLUID_REGISTRY
 } from "./types";
+import isUserLogin from "../securityUtils/isUserLogin";
 
 export const addFluidRegistry = (fluidRegistry) => async (dispatch) => {
 	try {
-		const res = await axios.post("/api/fluid-registries", fluidRegistry);
-		dispatch({
-			type: ADD_FLUID_REGISTRY,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.post("/api/fluid-registries", fluidRegistry);
+			dispatch({
+				type: ADD_FLUID_REGISTRY,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -28,30 +31,34 @@ export const addFluidRegistry = (fluidRegistry) => async (dispatch) => {
 };
 
 export const getFluidRegistries = () => async (dispatch) => {
-	const res = await axios.get("/api/fluid-registries");
-	dispatch({
-		type: GET_FLUID_REGISTRIES,
-		payload: res.data
-	});
+	if (isUserLogin()) {
+		const res = await axios.get("/api/fluid-registries");
+		dispatch({
+			type: GET_FLUID_REGISTRIES,
+			payload: res.data
+		});
+	}
 };
 
 export const updateFluidRegistry = (registryId, updatedfluidRegistry) => async (
 	dispatch
 ) => {
 	try {
-		const res = await axios.put(
-			`/api/fluid-registries/${registryId}`,
-			updatedfluidRegistry
-		);
-		dispatch({
-			type: UPDATE_FLUID_REGISTRY,
-			payload: res.data
-		});
-		dispatch({
-			type: GET_ERRORS,
-			payload: {}
-		});
-		return res;
+		if (isUserLogin()) {
+			const res = await axios.put(
+				`/api/fluid-registries/${registryId}`,
+				updatedfluidRegistry
+			);
+			dispatch({
+				type: UPDATE_FLUID_REGISTRY,
+				payload: res.data
+			});
+			dispatch({
+				type: GET_ERRORS,
+				payload: {}
+			});
+			return res;
+		}
 	} catch (error) {
 		dispatch({
 			type: GET_ERRORS,
@@ -61,12 +68,14 @@ export const updateFluidRegistry = (registryId, updatedfluidRegistry) => async (
 };
 
 export const deleteFluidRegistry = (registryId) => async (dispatch) => {
-	if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie wpisu.")) {
-		await axios.delete(`/api/fluid-registries/${registryId}`);
-		dispatch({
-			type: DELETE_FLUID_REGISTRY,
-			payload: registryId
-		});
+	if (isUserLogin()) {
+		if (window.confirm("Czy jesteś pewny? Spowoduje to usunięcie wpisu.")) {
+			await axios.delete(`/api/fluid-registries/${registryId}`);
+			dispatch({
+				type: DELETE_FLUID_REGISTRY,
+				payload: registryId
+			});
+		}
 	}
 };
